Lazily initialize Swift `LinkingObjects` instances.
`LinkingObjects` instances were being initialized eagerly by
`RLMInitializeSwiftAccessorGenerics`, which was resulting in a backlinks
`TableView` being created for every Swift instance of a type that
contained a `LinkingObjects` property. Lazily initializing these
properties means that the work of building the backlinks view is
deferred until the property is first accessed. This significantly
reduces the cost of allocating a Swift object that has a
`LinkingObjects` property.